requirements.txt
streamlit
seaborn
pandas
plotly
pillow
matplotlib
numpy
scikit-learn==1.6.1